Add first section and build system
Add the main .tex file, the first section and a Makefile to more easily
build the presentation.

As far as I know, LaTeX is quite awful with incremental builds (or, at the
very least, pdflatex is). However, using a Makefile still seems useful as
it's able to detect if any file has changed before rebuilding the slides.